Avoid modifying `require.paths`:
http://nodejs.org/docs/v0.4.4/api/modules.html -

**Note:** Please Avoid Modifying `require.paths`
For compatibility reasons, require.paths is still 
given first priority in the module lookup process.
However, it may disappear in a future release.

While it seemed like a good idea at the time, and
enabled a lot of useful experimentation, in
practice a mutable require.paths list is often a
troublesome source of confusion and headaches.