Add a watch option to the node.js compiler
By adding the argument '-w' or '--watch', lessc will use the fs.watchFile
method to watch for changes inside the directory the given less file
resides and recompile it on each change. Not the given file, but the
directory is used so changes to files which are included with a @import
statement trigger a recompile as well.
Only works properly if the source and destination arguments are given
and pointing to writeable / readable files