Fix #423 by preserving order of link/style elements
Appending generated css to the head element is confusing because it
changes the order of style declarations. If there are regular css link
or style elements after the less link element, less declarations with
equal specificity will override css declarations that come later in the
document.

This change preserves the original order by inserting the generated css
immediately after the less link element.