Added support for default variables
You can now define default variables via `@var: 1px !default;`. It will
behave as `@var: 1px;` if `@var` hasnt't been defined yet. Otherwise
it will be ignored.

This is useful for frameworks (e.g. Bootstrap) which need to define
variables to work but want to allow the user of the framework to
change them without messing with the internal less-stylesheets.