Update less-1.4.2.js
Hi, I had problems with Ajax connections to local files and I think i've already fix it. The changes are in lines 5714 and 5718. With these changes the connections of IE10 work without errors. It is the same problem that jQuery had in the past. Otherwise, perhaps you can use the Ajax connection from jQuery and won't have to fight with these things.

Greetings.