Check location.port for truthiness
According to the w3 spec for Window:

http://www.w3.org/TR/Window/#location

`location.port` might be null. This PR checks that it is truthy before checking for the port's length.

One place where location.port is null is within [jsdom](https://github.com/tmpvar/jsdom), so if you run less within Node, within jsdom, it throws an exception in this code.