variable variables with named colors error
I not sure if this PR make sense at all. But i found that:

```css
@color: blue;
@blue: yellow;
s{
p: @@color;
}
```

throws: `NameError: variable @undefined is undefined in`, the error makes sense cause a color can not be the name of a variable, but the error message does not.
Instead of trowing an error, the problems seems also to fix with `v.value.value = v.value.keyword;`.

Possible everybody knows already have to use only strings  or only the documentation at http://lesscss.org/features/#variables-feature-variable-names should be updated for that.