browser: Copy a Content-Security-Policy nonce attribute from <link>
When a nonce attribute is in use by Content-Security-Policy
copy the attribute from the <link> tag to the <style> tag.
This allows the Content-Security-Policy to accept the <style>
tag that's been created by the in-browser less.

The goal is for developers to be running with Content-Security-Policy
set to enforcing. This ensures that developers find bugs related to
the policy. But at the same time still having less process stylesheets
in the browser during development.